requirements.txt
discord.py
pymongo
pyyaml
typing_extensions
